Retructure project
