--- PlutodoApiServer/src/main/java/org/plutotramble/AsyncConfig.java
@@ -1,30 +1,0 @@
-package org.plutotramble;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.task.TaskExecutor;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-import org.springframework.security.task.DelegatingSecurityContextAsyncTaskExecutor;
-
-@Configuration
-public class AsyncConfig {
-
-    @Bean
-    public ThreadPoolTaskExecutor threadPoolTaskExecutor() {
-        final ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setThreadNamePrefix("PlutodoAsync-");
-        executor.setCorePoolSize(10);
-        executor.setMaxPoolSize(20);
-        executor.setQueueCapacity(100);
-        executor.setWaitForTasksToCompleteOnShutdown(true);
-        executor.setAwaitTerminationSeconds(60);
-        executor.initialize();
-        return executor;
-    }
-
-    @Bean
-    public TaskExecutor taskExecutor() {
-        return new DelegatingSecurityContextAsyncTaskExecutor(threadPoolTaskExecutor()) {
-        };
-    }
-}
