--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/dto/LoginDTO.java
@@ -1,6 +1,0 @@
-package org.plutotramble.authentication.dto;
-
-public class LoginDTO {
-    public String username = "";
-    public String password = "";
-}
