--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/dto/RegisterDTO.java
@@ -1,8 +1,0 @@
-package org.plutotramble.authentication.dto;
-
-public class RegisterDTO {
-    public String emailAddress = "";
-    public String username = "";
-    public String password = "";
-    public String passwordConfirm = "";
-}
