--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/exceptions/EmailAddressAlreadyExistsAuthenticationException.java
@@ -1,9 +1,0 @@
-package org.plutotramble.authentication.exceptions;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class EmailAddressAlreadyExistsAuthenticationException extends AuthenticationException {
-    public EmailAddressAlreadyExistsAuthenticationException(final String msg) {
-        super(msg);
-    }
-}
