--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/exceptions/InvalidEmailAddressException.java
@@ -1,9 +1,0 @@
-package org.plutotramble.authentication.exceptions;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidEmailAddressException extends AuthenticationException {
-    public InvalidEmailAddressException(final String msg) {
-        super(msg);
-    }
-}
