--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/exceptions/InvalidPasswordException.java
@@ -1,9 +1,0 @@
-package org.plutotramble.authentication.exceptions;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidPasswordException extends AuthenticationException {
-    public InvalidPasswordException(final String msg) {
-        super(msg);
-    }
-}
