--- PlutodoApiServer/src/main/java/org/plutotramble/authentication/exceptions/InvalidUsernameException.java
@@ -1,9 +1,0 @@
-package org.plutotramble.authentication.exceptions;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidUsernameException extends AuthenticationException {
-    public InvalidUsernameException(final String msg) {
-        super(msg);
-    }
-}
