--- PlutodoApiServer/src/main/java/org/plutotramble/category/CategoryDTO.java
@@ -1,11 +1,0 @@
-package org.plutotramble.category;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public class CategoryDTO {
-    public UUID id;
-    public String name;
-    public String color;
-    public LocalDateTime dateCreated;
-}
