+++ PlutodoApiServer/src/main/java/org/plutotramble/dto/CategoryDTO.java
@@ -1,0 +1,11 @@
+package org.plutotramble.dto;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+public class CategoryDTO {
+    public UUID id;
+    public String name;
+    public String color;
+    public LocalDateTime dateCreated;
+}
