+++ PlutodoApiServer/src/main/java/org/plutotramble/dto/LoginDTO.java
@@ -1,0 +1,6 @@
+package org.plutotramble.dto;
+
+public class LoginDTO {
+    public String username = "";
+    public String password = "";
+}
