+++ PlutodoApiServer/src/main/java/org/plutotramble/dto/RegisterDTO.java
@@ -1,0 +1,8 @@
+package org.plutotramble.dto;
+
+public class RegisterDTO {
+    public String emailAddress = "";
+    public String username = "";
+    public String password = "";
+    public String passwordConfirm = "";
+}
