+++ PlutodoApiServer/src/main/java/org/plutotramble/entity/CalendarEventEntity.java
@@ -1,0 +1,114 @@
+package org.plutotramble.entity;
+
+import jakarta.persistence.*;
+
+import java.sql.Timestamp;
+import java.util.Objects;
+import java.util.UUID;
+
+@Entity
+@Table(name = "calendar_event", schema = "public", catalog = "javadevplutodo")
+public class CalendarEventEntity {
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @Column(name = "id", nullable = false)
+    private UUID id;
+    @Basic
+    @Column(name = "name", nullable = false, length = 30)
+    private String name;
+    @Basic
+    @Column(name = "description", nullable = true, length = 8192)
+    private String description;
+    @Basic
+    @Column(name = "date_created", nullable = false)
+    private Timestamp dateCreated;
+    @Basic
+    @Column(name = "date_from", nullable = false)
+    private Timestamp dateFrom;
+    @Basic
+    @Column(name = "date_to", nullable = true)
+    private Timestamp dateTo;
+    @Basic
+    @Column(name = "is_whole_day", nullable = false)
+    private boolean isWholeDay;
+    @ManyToOne
+    @JoinColumn(name = "category_id", referencedColumnName = "id", nullable = false)
+    private CategoryEntity category;
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Timestamp getDateCreated() {
+        return dateCreated;
+    }
+
+    public void setDateCreated(Timestamp dateCreated) {
+        this.dateCreated = dateCreated;
+    }
+
+    public Timestamp getDateFrom() {
+        return dateFrom;
+    }
+
+    public void setDateFrom(Timestamp dateFrom) {
+        this.dateFrom = dateFrom;
+    }
+
+    public Timestamp getDateTo() {
+        return dateTo;
+    }
+
+    public void setDateTo(Timestamp dateTo) {
+        this.dateTo = dateTo;
+    }
+
+    public boolean isWholeDay() {
+        return isWholeDay;
+    }
+
+    public void setWholeDay(boolean wholeDay) {
+        isWholeDay = wholeDay;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        CalendarEventEntity that = (CalendarEventEntity) o;
+        return isWholeDay == that.isWholeDay && Objects.equals(id, that.id) && Objects.equals(name, that.name) && Objects.equals(description, that.description) && Objects.equals(dateCreated, that.dateCreated) && Objects.equals(dateFrom, that.dateFrom) && Objects.equals(dateTo, that.dateTo);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, description, dateCreated, dateFrom, dateTo, isWholeDay);
+    }
+
+    public CategoryEntity getCategory() {
+        return category;
+    }
+
+    public void setCategory(CategoryEntity category) {
+        this.category = category;
+    }
+}
