+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/ErrorMessage.java
@@ -1,0 +1,8 @@
+package org.plutotramble.exception;
+
+public class ErrorMessage {
+    public ErrorMessage(String message){
+        error = message;
+    }
+    String error;
+}
