+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/InvalidItemPropertyException.java
@@ -1,0 +1,7 @@
+package org.plutotramble.exception;
+
+public class InvalidItemPropertyException extends Exception {
+    public InvalidItemPropertyException(String msg){
+        super(msg);
+    }
+}
