+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/ItemNotFoundException.java
@@ -1,0 +1,7 @@
+package org.plutotramble.exception;
+
+public class ItemNotFoundException extends Exception {
+    public ItemNotFoundException(String msg) {
+        super(msg);
+    }
+}
