+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/authentication/EmailAddressAlreadyExistsAuthenticationException.java
@@ -1,0 +1,9 @@
+package org.plutotramble.exception.authentication;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class EmailAddressAlreadyExistsAuthenticationException extends AuthenticationException {
+    public EmailAddressAlreadyExistsAuthenticationException(final String msg) {
+        super(msg);
+    }
+}
