+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/authentication/InvalidEmailAddressException.java
@@ -1,0 +1,9 @@
+package org.plutotramble.exception.authentication;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class InvalidEmailAddressException extends AuthenticationException {
+    public InvalidEmailAddressException(final String msg) {
+        super(msg);
+    }
+}
