+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/authentication/InvalidPasswordException.java
@@ -1,0 +1,9 @@
+package org.plutotramble.exception.authentication;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class InvalidPasswordException extends AuthenticationException {
+    public InvalidPasswordException(final String msg) {
+        super(msg);
+    }
+}
