+++ PlutodoApiServer/src/main/java/org/plutotramble/exception/authentication/InvalidUsernameException.java
@@ -1,0 +1,9 @@
+package org.plutotramble.exception.authentication;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class InvalidUsernameException extends AuthenticationException {
+    public InvalidUsernameException(final String msg) {
+        super(msg);
+    }
+}
