--- PlutodoApiServer/src/main/java/org/plutotramble/shared/exceptions/ErrorMessage.java
@@ -1,8 +1,0 @@
-package org.plutotramble.shared.exceptions;
-
-public class ErrorMessage {
-    public ErrorMessage(String message){
-        error = message;
-    }
-    String error;
-}
