--- PlutodoApiServer/src/main/java/org/plutotramble/shared/exceptions/InvalidItemPropertyException.java
@@ -1,7 +1,0 @@
-package org.plutotramble.shared.exceptions;
-
-public class InvalidItemPropertyException extends Exception {
-    public InvalidItemPropertyException(String msg){
-        super(msg);
-    }
-}
