--- PlutodoApiServer/src/main/java/org/plutotramble/shared/exceptions/ItemNotFoundException.java
@@ -1,7 +1,0 @@
-package org.plutotramble.shared.exceptions;
-
-public class ItemNotFoundException extends Exception {
-    public ItemNotFoundException(String msg) {
-        super(msg);
-    }
-}
