--- PlutodoApiServer/src/main/java/org/plutotramble/task/TaskItemDTO.java
@@ -1,10 +1,0 @@
-package org.plutotramble.task;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public class TaskItemDTO extends SmallTaskItemDTO {
-    public String description;
-    public LocalDateTime dateCreated;
-    public UUID categoryId;
-}
